Show MIME-type instead of useless friendly name
When the friendly MIME type for the guessed mime type is
unavailable, do not show the friendly name for the real MIME-type,
because it's confusing. For persistent settings, the dialog will
always use the guessed MIME type, so showing "unknown" instead of
"application/something-weird" is not helpful.

Note that the guessed MIME-type only differs from the real MIME-type
when the real MIME-type is either application/octet-stream or
text/plain. See background.js for this logic.