Use a different way to find frames
chrome.webNavigation.getAllFrames returns the URL of the last
navigation, even if the URL has been changed.

chrome.contextMenus.onClicked returns the current URL of the
document, which may be different from the last navigated URL.

Because of this inconsistency, I decided to switch to a different
method using document.hasFocus.