Distinguish server-sent MIME from sniffed MIME
Previously, if the user opts in to permanently using a certain option
for server-sent MIME-types, then the choice would be saved for the
sniffed MIME instead of the original MIME.

Then we could have the odd situation where "application/pdf"'s action
is set to opening as "application/octet-stream".

Now we distinguish the two options. In the preferences there is a
separate table for sniffed MIME types (with fall back to server-sent
MIME).