Remove contextmenu support part 1/2
I initially introduced a context menu because Chrome did not have a
reliable way to intercept requests (before I added redirectUrl support
at the onHeadersReceived event in the webRequest API).
Chrome has supported redirectUrl for years now, so that reason is gone.

Since I don't use the context menu, and there were no feature requests
to support it in the original repo, I'll remove it.

Note that the original add-on claims to support a "View > View as" menu
item, but it has not worked for a long while, with little complaints:
https://github.com/spasche/openinbrowser/issues/27
So probably nobody will miss the feature.