Remove support for "extensions" as an external viewer.
It is not possible to obtain a useful moz-extension URL, not even
through the management API due to the random GUID in the host name.
So, remove it.

The only example of an "extension" is the PDF.js Chrome/Opera extension.
In Firefox PDF.js is used by default, so there is no need to detect that
extension.