Correctly center dialog on multi-display setups (#6)
Credits to @davidf-woaf-net for reporting the issue and proposing the
solution.

Note that `screen.left` and `screen.top` are non-standard properties.
They are undefined on Chrome, and trying to use it would result in
attempting to call `windows.update` with `NaN` as a value. This is
OK, if I ever decide to port the extension to Chrome, then this
serves as a reminder that I need to test the extension on multi-display
setups and adjust the code as necessary.