Fix: Always guess MIME from file name
In 677834f441a22b6d2f2a81a9edf531e1914fc5e0, the "MIME from file
extension" option started to always be displayed. So the file name
should always be guessed, not only if `isSniffingMimeType` is set.

Now that the MIME is always guessed, there is no compelling need for
forcing a different MIME for text/plain responses, especially since
text/plain is either inline or offered as a download.
So now `isSniffingMimeType` is only set for `application/octet-stream`.