Add more mime icons, including duplicates
The png files in extensions/icons/ are used to generate the
"nonGenericIcons" mapping in mime-metadata.js. Previously I omited
symlinks because I assumed that the images would be a duplicate of
another icon in the icons directory.

This is not the case; some icons are symlinked to a file outside of
/usr/share/icons/oxygen/base/32x32/mimetypes.

I have now copied all images using:

   for f in $(find /usr/share/icons/oxygen/base/32x32/mimetypes/ -name '*.png' -exec basename {} \;); do rsync -L "/usr/share/icons/oxygen/base/32x32/mimetypes/$f" "$f" ; done

Total file size:
$ du -cbs *.png | tail -n1
706369  total

Total file size minus duplicates:
$ du -cbs $(md5sum *.png | sort | uniq -w32 | cut -d' ' -f3) | tail -n1
345732  total

So the file size with duplicates is about twice as much as without.
This can be optimized later.