Use mime_getIcon for action detection
The previous method did not take into account
`mimeMetadata.nonGenericIcons` (or `mimeMetadata.aliases`).
Re-use `mime_getIcon` to make use of this, so that the mapping is
correct.

This fixes an issue where "text/html" is treated as "text-x-generic"
instead of "text-html".

To test: Tools menu > Open in Browser > "Enabled for next request",
reload any HTML page and confirm that the default action is
"Open in browser as web page".