Use browser's font for dialog + add vertical space (#21)
Let the browser decide which font and size to use.
Also added some vertical whitespace to avoid a crammed appearance.

The "message-box" and "menu" were chosen as values for the "font" CSS
property because inspecting these XUL elements in the dialog using the
browser toolbox showed that these values are actually in use.
This should ensure that the font and size of the custom dialog is
identical to the built-in dialog.