Use server-sent MIME when "Do this automatically" is used (#58)
When the "Do this automatically" checkbox is checked, the chosen value
should be associated with the server-sent MIME type, because that is
used later for looking up the previously chosen value.

The only exception is when the user has enabled the option
"Get MIME from file extension for application/octet-stream".
In that case, the sniffed MIME type may be used instead.